requirements.txt
pycryptodome==3.19.1
matplotlib==3.8.3
flask==3.0.2 